Adding Landfill Support
Adds landfill support, plus a lot of backend work to start supporting the new config GUI.
