+++ concreep-redux/data.lua
@@ -1,0 +1,53 @@
+data:extend({
+	{
+		type = "custom-input",
+		name = "concreep_toggle_interface",
+		key_sequence = "CONTROL + J",
+		order = "a"
+	}
+})
+
+local styles = data.raw["gui-style"].default
+
+styles["ccr_content_frame"] = {
+	type = "frame_style",
+	parent = "inside_shallow_frame_with_padding",
+	vertically_stretchable = "on"
+}
+
+styles["ccr_controls_flow"] = {
+	type = "horizontal_flow_style",
+	vertical_align = "center",
+	horizontal_spacing = 16
+}
+
+styles["ccr_controls_textfield"] = {
+	type = "textbox_style",
+	width = 36
+}
+
+styles["ccr_tile_list_row"] = {
+	type = "frame_style",
+	parent = "dark_frame",
+	horizontally_stretchable = "on",
+}
+
+styles["ccr_tile_list_draghandle"] = {
+	type = "empty_widget_style",
+	parent = "draggable_space",
+	width = 75,
+	vertically_stretchable = "on",
+	horizontally_stretchable = "off",
+	horizontal_align="right"
+}
+
+styles["ccr_deep_frame"] = {
+	type = "frame_style",
+	parent = "slot_button_deep_frame",
+	vertically_stretchable = "on",
+	horizontally_stretchable = "on",
+	top_margin = 16,
+	left_margin = 8,
+	right_margin = 8,
+	bottom_margin = 4
+}
