+++ concreep-redux/locale/en/locale.cfg
@@ -10,6 +10,8 @@ upgrade-space-scaffold=Upgrade Space Scaffold
 concreep-tiles-per-area=Tiles Per Coverage Type
 concreep-logistic-area-tile=Tile for Logistic Range
 concreep-construction-area-tile=Tile for Construction Range
+creep-landfill=Place Landfill
+concreep-pump-radius=Landfill Pump Clearance
 
 [mod-setting-description]
 concreep-range=Range that roboports will automatically place concreep as a percentage of roboport construction range.
@@ -21,3 +23,5 @@ upgrade-brick=Upgrade brick tiles to concrete or reinforced concrete. Turn off i
 upgrade-concrete=Upgrade concrete tiles to reinforced concrete. Turn off if you want to use concrete accents for decoration.
 upgrade-space-scaffold=Upgrade space scaffolding to space plating. Has no effect if Space Exploration isn't loaded.
 concreep-tiles-per-area=Enable selection of different tile types to use based on whether a given location is in logistic or construction range of a roboport. Note: This disables the normal 'upgrade' logic, and just tries to place the select types. Construction won't overwrite logistic.
+creep-landfill=If water is encountered, place landfill.
+concreep-pump-radius=How much clearance to leave around a pump when placing landfill.
